Finalizando aplicación
Mejorando alerts, mensaje en tabla vacia mis rutas, problema de
visualización al pulsar en imprimir...

+++ greatRouteAPP/src/main/java/com/greatRoute/services/impl/RutaServiceImpl.java
@@ -65,8 +65,10 @@ public class RutaServiceImpl implements RutaService {
 	@Override
 	public void borrarRuta(int id) {
 		Ruta ruta = rutaRepository.findById(id);
-		File file = new File(ruta.getRecorrido());
-		file.delete();
+		if(ruta!=null){
+			File file = new File(ruta.getRecorrido());
+			file.delete();
+		}
 		rutaRepository.delete(ruta);
 	}
